+++ tiledesk-client/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Client library for Tiledesk nodejs APIs client
 
+### 0.7.7 - online
+- bug fix: APIURL
+
 ### 0.7.6 - online
 - added updateProjectUserCurrentlyLoggedIn
 - added "options" as last parameter of getRequests to allow adding additional params as no_populate, snap_department etc.